requirements.txt
alembic==1.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SocketIO==3.0.2
Flask-SQLAlchemy==2.4.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysql-connector==2.2.9
pyserial==3.4
python-dateutil==2.7.3
python-editor==1.0.3
python-engineio==2.3.2
python-socketio==2.0.0
requests==2.20.0
six==1.11.0
SQLAlchemy==1.3.3
urllib3==1.23
Werkzeug==0.14.1
